requirements.txt
astroid==2.3.3
beautifulsoup4==4.9.0
billiard==3.6.3.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
pyaes==1.6.1
pylint==2.4.4
Pyrogram==0.16.0
PySocks==1.7.0
python-decouple==3.3
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
TgCrypto==1.2.0
typed-ast==1.4.1
urllib3==1.25.9
vine==1.3.0
wrapt==1.11.2
zipp==3.1.0
